@@ -1,19 +1,25 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/sarchlab/mgpusim/v4/amd/benchmarks/dnn/layers (interfaces: Layer)
+//
+// Generated by this command:
+//
+//	mockgen -destination mock_layers_test.go -package training -write_package_comment=false github.com/sarchlab/mgpusim/v4/amd/benchmarks/dnn/layers Layer
+//
 
 package training
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	tensor "github.com/sarchlab/mgpusim/v4/amd/benchmarks/dnn/tensor"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockLayer is a mock of Layer interface.
 type MockLayer struct {
 	ctrl     *gomock.Controller
 	recorder *MockLayerMockRecorder
+	isgomock struct{}
 }
 
 // MockLayerMockRecorder is the mock recorder for MockLayer.
@@ -34,31 +40,31 @@ func (m *MockLayer) EXPECT() *MockLayerMockRecorder {
 }
 
 // Backward mocks base method.
-func (m *MockLayer) Backward(arg0 tensor.Tensor) tensor.Tensor {
+func (m *MockLayer) Backward(input tensor.Tensor) tensor.Tensor {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Backward", arg0)
+	ret := m.ctrl.Call(m, "Backward", input)
 	ret0, _ := ret[0].(tensor.Tensor)
 	return ret0
 }
 
 // Backward indicates an expected call of Backward.
-func (mr *MockLayerMockRecorder) Backward(arg0 interface{}) *gomock.Call {
+func (mr *MockLayerMockRecorder) Backward(input any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Backward", reflect.TypeOf((*MockLayer)(nil).Backward), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Backward", reflect.TypeOf((*MockLayer)(nil).Backward), input)
 }
 
 // Forward mocks base method.
-func (m *MockLayer) Forward(arg0 tensor.Tensor) tensor.Tensor {
+func (m *MockLayer) Forward(input tensor.Tensor) tensor.Tensor {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Forward", arg0)
+	ret := m.ctrl.Call(m, "Forward", input)
 	ret0, _ := ret[0].(tensor.Tensor)
 	return ret0
 }
 
 // Forward indicates an expected call of Forward.
-func (mr *MockLayerMockRecorder) Forward(arg0 interface{}) *gomock.Call {
+func (mr *MockLayerMockRecorder) Forward(input any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Forward", reflect.TypeOf((*MockLayer)(nil).Forward), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Forward", reflect.TypeOf((*MockLayer)(nil).Forward), input)
 }
 
 // Gradients mocks base method.
