@@ -1,19 +1,25 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/sarchlab/mgpusim/v4/amd/benchmarks/dnn/training/optimization (interfaces: Alg)
+//
+// Generated by this command:
+//
+//	mockgen -destination mock_optimization_test.go -package training -write_package_comment=false github.com/sarchlab/mgpusim/v4/amd/benchmarks/dnn/training/optimization Alg
+//
 
 package training
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	optimization "github.com/sarchlab/mgpusim/v4/amd/benchmarks/dnn/training/optimization"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockAlg is a mock of Alg interface.
 type MockAlg struct {
 	ctrl     *gomock.Controller
 	recorder *MockAlgMockRecorder
+	isgomock struct{}
 }
 
 // MockAlgMockRecorder is the mock recorder for MockAlg.
@@ -34,13 +40,13 @@ func (m *MockAlg) EXPECT() *MockAlgMockRecorder {
 }
 
 // UpdateParameters mocks base method.
-func (m *MockAlg) UpdateParameters(arg0 optimization.Layer) {
+func (m *MockAlg) UpdateParameters(l optimization.Layer) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "UpdateParameters", arg0)
+	m.ctrl.Call(m, "UpdateParameters", l)
 }
 
 // UpdateParameters indicates an expected call of UpdateParameters.
-func (mr *MockAlgMockRecorder) UpdateParameters(arg0 interface{}) *gomock.Call {
+func (mr *MockAlgMockRecorder) UpdateParameters(l any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpdateParameters", reflect.TypeOf((*MockAlg)(nil).UpdateParameters), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpdateParameters", reflect.TypeOf((*MockAlg)(nil).UpdateParameters), l)
 }
