@@ -1,19 +1,25 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/sarchlab/mgpusim/v4/amd/benchmarks/dnn/training (interfaces: LossFunction,DataSource)
+//
+// Generated by this command:
+//
+//	mockgen -destination mock_training_test.go -package training -self_package training -write_package_comment=false github.com/sarchlab/mgpusim/v4/amd/benchmarks/dnn/training LossFunction,DataSource
+//
 
 package training
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	tensor "github.com/sarchlab/mgpusim/v4/amd/benchmarks/dnn/tensor"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockLossFunction is a mock of LossFunction interface.
 type MockLossFunction struct {
 	ctrl     *gomock.Controller
 	recorder *MockLossFunctionMockRecorder
+	isgomock struct{}
 }
 
 // MockLossFunctionMockRecorder is the mock recorder for MockLossFunction.
@@ -34,24 +40,25 @@ func (m *MockLossFunction) EXPECT() *MockLossFunctionMockRecorder {
 }
 
 // Loss mocks base method.
-func (m *MockLossFunction) Loss(arg0 tensor.Tensor, arg1 []int) (float64, tensor.Tensor) {
+func (m *MockLossFunction) Loss(output tensor.Tensor, label []int) (float64, tensor.Tensor) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Loss", arg0, arg1)
+	ret := m.ctrl.Call(m, "Loss", output, label)
 	ret0, _ := ret[0].(float64)
 	ret1, _ := ret[1].(tensor.Tensor)
 	return ret0, ret1
 }
 
 // Loss indicates an expected call of Loss.
-func (mr *MockLossFunctionMockRecorder) Loss(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockLossFunctionMockRecorder) Loss(output, label any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Loss", reflect.TypeOf((*MockLossFunction)(nil).Loss), arg0, arg1)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Loss", reflect.TypeOf((*MockLossFunction)(nil).Loss), output, label)
 }
 
 // MockDataSource is a mock of DataSource interface.
 type MockDataSource struct {
 	ctrl     *gomock.Controller
 	recorder *MockDataSourceMockRecorder
+	isgomock struct{}
 }
 
 // MockDataSourceMockRecorder is the mock recorder for MockDataSource.
@@ -72,18 +79,18 @@ func (m *MockDataSource) EXPECT() *MockDataSourceMockRecorder {
 }
 
 // NextBatch mocks base method.
-func (m *MockDataSource) NextBatch(arg0 int) (tensor.Tensor, []int) {
+func (m *MockDataSource) NextBatch(batchSize int) (tensor.Tensor, []int) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "NextBatch", arg0)
+	ret := m.ctrl.Call(m, "NextBatch", batchSize)
 	ret0, _ := ret[0].(tensor.Tensor)
 	ret1, _ := ret[1].([]int)
 	return ret0, ret1
 }
 
 // NextBatch indicates an expected call of NextBatch.
-func (mr *MockDataSourceMockRecorder) NextBatch(arg0 interface{}) *gomock.Call {
+func (mr *MockDataSourceMockRecorder) NextBatch(batchSize any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NextBatch", reflect.TypeOf((*MockDataSource)(nil).NextBatch), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NextBatch", reflect.TypeOf((*MockDataSource)(nil).NextBatch), batchSize)
 }
 
 // Rewind mocks base method.
