@@ -1,19 +1,25 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/sarchlab/akita/v4/mem/vm (interfaces: PageTable)
+//
+// Generated by this command:
+//
+//	mockgen -destination mock_vm_test.go -package emu -write_package_comment=false github.com/sarchlab/akita/v4/mem/vm PageTable
+//
 
 package emu
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	vm "github.com/sarchlab/akita/v4/mem/vm"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockPageTable is a mock of PageTable interface.
 type MockPageTable struct {
 	ctrl     *gomock.Controller
 	recorder *MockPageTableMockRecorder
+	isgomock struct{}
 }
 
 // MockPageTableMockRecorder is the mock recorder for MockPageTable.
@@ -34,52 +40,52 @@ func (m *MockPageTable) EXPECT() *MockPageTableMockRecorder {
 }
 
 // Find mocks base method.
-func (m *MockPageTable) Find(arg0 vm.PID, arg1 uint64) (vm.Page, bool) {
+func (m *MockPageTable) Find(pid vm.PID, Addr uint64) (vm.Page, bool) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Find", arg0, arg1)
+	ret := m.ctrl.Call(m, "Find", pid, Addr)
 	ret0, _ := ret[0].(vm.Page)
 	ret1, _ := ret[1].(bool)
 	return ret0, ret1
 }
 
 // Find indicates an expected call of Find.
-func (mr *MockPageTableMockRecorder) Find(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockPageTableMockRecorder) Find(pid, Addr any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Find", reflect.TypeOf((*MockPageTable)(nil).Find), arg0, arg1)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Find", reflect.TypeOf((*MockPageTable)(nil).Find), pid, Addr)
 }
 
 // Insert mocks base method.
-func (m *MockPageTable) Insert(arg0 vm.Page) {
+func (m *MockPageTable) Insert(page vm.Page) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "Insert", arg0)
+	m.ctrl.Call(m, "Insert", page)
 }
 
 // Insert indicates an expected call of Insert.
-func (mr *MockPageTableMockRecorder) Insert(arg0 interface{}) *gomock.Call {
+func (mr *MockPageTableMockRecorder) Insert(page any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Insert", reflect.TypeOf((*MockPageTable)(nil).Insert), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Insert", reflect.TypeOf((*MockPageTable)(nil).Insert), page)
 }
 
 // Remove mocks base method.
-func (m *MockPageTable) Remove(arg0 vm.PID, arg1 uint64) {
+func (m *MockPageTable) Remove(pid vm.PID, vAddr uint64) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "Remove", arg0, arg1)
+	m.ctrl.Call(m, "Remove", pid, vAddr)
 }
 
 // Remove indicates an expected call of Remove.
-func (mr *MockPageTableMockRecorder) Remove(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockPageTableMockRecorder) Remove(pid, vAddr any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Remove", reflect.TypeOf((*MockPageTable)(nil).Remove), arg0, arg1)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Remove", reflect.TypeOf((*MockPageTable)(nil).Remove), pid, vAddr)
 }
 
 // Update mocks base method.
-func (m *MockPageTable) Update(arg0 vm.Page) {
+func (m *MockPageTable) Update(page vm.Page) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "Update", arg0)
+	m.ctrl.Call(m, "Update", page)
 }
 
 // Update indicates an expected call of Update.
-func (mr *MockPageTableMockRecorder) Update(arg0 interface{}) *gomock.Call {
+func (mr *MockPageTableMockRecorder) Update(page any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Update", reflect.TypeOf((*MockPageTable)(nil).Update), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Update", reflect.TypeOf((*MockPageTable)(nil).Update), page)
 }
