@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: alg.go
+//
+// Generated by this command:
+//
+//	mockgen -source alg.go -destination mock_alg.go -package dispatching -mock_names=algorithm=MockAlgorithm
+//
 
 // Package dispatching is a generated GoMock package.
 package dispatching
@@ -7,15 +12,16 @@ package dispatching
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	kernels "github.com/sarchlab/mgpusim/v4/amd/kernels"
 	resource "github.com/sarchlab/mgpusim/v4/amd/timing/cp/internal/resource"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockAlgorithm is a mock of algorithm interface.
 type MockAlgorithm struct {
 	ctrl     *gomock.Controller
 	recorder *MockAlgorithmMockRecorder
+	isgomock struct{}
 }
 
 // MockAlgorithmMockRecorder is the mock recorder for MockAlgorithm.
@@ -42,7 +48,7 @@ func (m *MockAlgorithm) FreeResources(location dispatchLocation) {
 }
 
 // FreeResources indicates an expected call of FreeResources.
-func (mr *MockAlgorithmMockRecorder) FreeResources(location interface{}) *gomock.Call {
+func (mr *MockAlgorithmMockRecorder) FreeResources(location any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "FreeResources", reflect.TypeOf((*MockAlgorithm)(nil).FreeResources), location)
 }
@@ -96,7 +102,7 @@ func (m *MockAlgorithm) RegisterCU(cu resource.DispatchableCU) {
 }
 
 // RegisterCU indicates an expected call of RegisterCU.
-func (mr *MockAlgorithmMockRecorder) RegisterCU(cu interface{}) *gomock.Call {
+func (mr *MockAlgorithmMockRecorder) RegisterCU(cu any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RegisterCU", reflect.TypeOf((*MockAlgorithm)(nil).RegisterCU), cu)
 }
@@ -108,7 +114,7 @@ func (m *MockAlgorithm) StartNewKernel(info kernels.KernelLaunchInfo) {
 }
 
 // StartNewKernel indicates an expected call of StartNewKernel.
-func (mr *MockAlgorithmMockRecorder) StartNewKernel(info interface{}) *gomock.Call {
+func (mr *MockAlgorithmMockRecorder) StartNewKernel(info any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "StartNewKernel", reflect.TypeOf((*MockAlgorithm)(nil).StartNewKernel), info)
 }
