@@ -1,19 +1,25 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/sarchlab/mgpusim/v4/amd/kernels (interfaces: GridBuilder)
+//
+// Generated by this command:
+//
+//	mockgen -destination mock_kernels_test.go -package dispatching -write_package_comment=false github.com/sarchlab/mgpusim/v4/amd/kernels GridBuilder
+//
 
 package dispatching
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	kernels "github.com/sarchlab/mgpusim/v4/amd/kernels"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockGridBuilder is a mock of GridBuilder interface.
 type MockGridBuilder struct {
 	ctrl     *gomock.Controller
 	recorder *MockGridBuilderMockRecorder
+	isgomock struct{}
 }
 
 // MockGridBuilderMockRecorder is the mock recorder for MockGridBuilder.
@@ -62,25 +68,25 @@ func (mr *MockGridBuilderMockRecorder) NumWG() *gomock.Call {
 }
 
 // SetKernel mocks base method.
-func (m *MockGridBuilder) SetKernel(arg0 kernels.KernelLaunchInfo) {
+func (m *MockGridBuilder) SetKernel(info kernels.KernelLaunchInfo) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "SetKernel", arg0)
+	m.ctrl.Call(m, "SetKernel", info)
 }
 
 // SetKernel indicates an expected call of SetKernel.
-func (mr *MockGridBuilderMockRecorder) SetKernel(arg0 interface{}) *gomock.Call {
+func (mr *MockGridBuilderMockRecorder) SetKernel(info any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetKernel", reflect.TypeOf((*MockGridBuilder)(nil).SetKernel), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetKernel", reflect.TypeOf((*MockGridBuilder)(nil).SetKernel), info)
 }
 
 // Skip mocks base method.
-func (m *MockGridBuilder) Skip(arg0 int) {
+func (m *MockGridBuilder) Skip(n int) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "Skip", arg0)
+	m.ctrl.Call(m, "Skip", n)
 }
 
 // Skip indicates an expected call of Skip.
-func (mr *MockGridBuilderMockRecorder) Skip(arg0 interface{}) *gomock.Call {
+func (mr *MockGridBuilderMockRecorder) Skip(n any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Skip", reflect.TypeOf((*MockGridBuilder)(nil).Skip), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Skip", reflect.TypeOf((*MockGridBuilder)(nil).Skip), n)
 }
