@@ -1,21 +1,27 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/sarchlab/mgpusim/v4/amd/timing/cp/internal/resource (interfaces: CUResourcePool,CUResource)
+//
+// Generated by this command:
+//
+//	mockgen -destination mock_resource_test.go -package dispatching -write_package_comment=false github.com/sarchlab/mgpusim/v4/amd/timing/cp/internal/resource CUResourcePool,CUResource
+//
 
 package dispatching
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	sim "github.com/sarchlab/akita/v4/sim"
 	kernels "github.com/sarchlab/mgpusim/v4/amd/kernels"
 	resource "github.com/sarchlab/mgpusim/v4/amd/timing/cp/internal/resource"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockCUResourcePool is a mock of CUResourcePool interface.
 type MockCUResourcePool struct {
 	ctrl     *gomock.Controller
 	recorder *MockCUResourcePoolMockRecorder
+	isgomock struct{}
 }
 
 // MockCUResourcePoolMockRecorder is the mock recorder for MockCUResourcePool.
@@ -36,17 +42,17 @@ func (m *MockCUResourcePool) EXPECT() *MockCUResourcePoolMockRecorder {
 }
 
 // GetCU mocks base method.
-func (m *MockCUResourcePool) GetCU(arg0 int) resource.CUResource {
+func (m *MockCUResourcePool) GetCU(i int) resource.CUResource {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetCU", arg0)
+	ret := m.ctrl.Call(m, "GetCU", i)
 	ret0, _ := ret[0].(resource.CUResource)
 	return ret0
 }
 
 // GetCU indicates an expected call of GetCU.
-func (mr *MockCUResourcePoolMockRecorder) GetCU(arg0 interface{}) *gomock.Call {
+func (mr *MockCUResourcePoolMockRecorder) GetCU(i any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetCU", reflect.TypeOf((*MockCUResourcePool)(nil).GetCU), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetCU", reflect.TypeOf((*MockCUResourcePool)(nil).GetCU), i)
 }
 
 // NumCU mocks base method.
@@ -64,21 +70,22 @@ func (mr *MockCUResourcePoolMockRecorder) NumCU() *gomock.Call {
 }
 
 // RegisterCU mocks base method.
-func (m *MockCUResourcePool) RegisterCU(arg0 resource.DispatchableCU) {
+func (m *MockCUResourcePool) RegisterCU(cu resource.DispatchableCU) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "RegisterCU", arg0)
+	m.ctrl.Call(m, "RegisterCU", cu)
 }
 
 // RegisterCU indicates an expected call of RegisterCU.
-func (mr *MockCUResourcePoolMockRecorder) RegisterCU(arg0 interface{}) *gomock.Call {
+func (mr *MockCUResourcePoolMockRecorder) RegisterCU(cu any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RegisterCU", reflect.TypeOf((*MockCUResourcePool)(nil).RegisterCU), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RegisterCU", reflect.TypeOf((*MockCUResourcePool)(nil).RegisterCU), cu)
 }
 
 // MockCUResource is a mock of CUResource interface.
 type MockCUResource struct {
 	ctrl     *gomock.Controller
 	recorder *MockCUResourceMockRecorder
+	isgomock struct{}
 }
 
 // MockCUResourceMockRecorder is the mock recorder for MockCUResource.
@@ -99,10 +106,10 @@ func (m *MockCUResource) EXPECT() *MockCUResourceMockRecorder {
 }
 
 // DispatchingPort mocks base method.
-func (m *MockCUResource) DispatchingPort() sim.Port {
+func (m *MockCUResource) DispatchingPort() sim.RemotePort {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "DispatchingPort")
-	ret0, _ := ret[0].(sim.Port)
+	ret0, _ := ret[0].(sim.RemotePort)
 	return ret0
 }
 
@@ -113,28 +120,28 @@ func (mr *MockCUResourceMockRecorder) DispatchingPort() *gomock.Call {
 }
 
 // FreeResourcesForWG mocks base method.
-func (m *MockCUResource) FreeResourcesForWG(arg0 *kernels.WorkGroup) {
+func (m *MockCUResource) FreeResourcesForWG(wg *kernels.WorkGroup) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "FreeResourcesForWG", arg0)
+	m.ctrl.Call(m, "FreeResourcesForWG", wg)
 }
 
 // FreeResourcesForWG indicates an expected call of FreeResourcesForWG.
-func (mr *MockCUResourceMockRecorder) FreeResourcesForWG(arg0 interface{}) *gomock.Call {
+func (mr *MockCUResourceMockRecorder) FreeResourcesForWG(wg any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "FreeResourcesForWG", reflect.TypeOf((*MockCUResource)(nil).FreeResourcesForWG), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "FreeResourcesForWG", reflect.TypeOf((*MockCUResource)(nil).FreeResourcesForWG), wg)
 }
 
 // ReserveResourceForWG mocks base method.
-func (m *MockCUResource) ReserveResourceForWG(arg0 *kernels.WorkGroup) ([]resource.WfLocation, bool) {
+func (m *MockCUResource) ReserveResourceForWG(wg *kernels.WorkGroup) ([]resource.WfLocation, bool) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ReserveResourceForWG", arg0)
+	ret := m.ctrl.Call(m, "ReserveResourceForWG", wg)
 	ret0, _ := ret[0].([]resource.WfLocation)
 	ret1, _ := ret[1].(bool)
 	return ret0, ret1
 }
 
 // ReserveResourceForWG indicates an expected call of ReserveResourceForWG.
-func (mr *MockCUResourceMockRecorder) ReserveResourceForWG(arg0 interface{}) *gomock.Call {
+func (mr *MockCUResourceMockRecorder) ReserveResourceForWG(wg any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ReserveResourceForWG", reflect.TypeOf((*MockCUResource)(nil).ReserveResourceForWG), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ReserveResourceForWG", reflect.TypeOf((*MockCUResource)(nil).ReserveResourceForWG), wg)
 }
