@@ -1,19 +1,25 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/sarchlab/akita/v4/tracing (interfaces: NamedHookable)
+//
+// Generated by this command:
+//
+//	mockgen -destination mock_tracing_test.go -package dispatching -write_package_comment=false github.com/sarchlab/akita/v4/tracing NamedHookable
+//
 
 package dispatching
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	sim "github.com/sarchlab/akita/v4/sim"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockNamedHookable is a mock of NamedHookable interface.
 type MockNamedHookable struct {
 	ctrl     *gomock.Controller
 	recorder *MockNamedHookableMockRecorder
+	isgomock struct{}
 }
 
 // MockNamedHookableMockRecorder is the mock recorder for MockNamedHookable.
@@ -34,15 +40,15 @@ func (m *MockNamedHookable) EXPECT() *MockNamedHookableMockRecorder {
 }
 
 // AcceptHook mocks base method.
-func (m *MockNamedHookable) AcceptHook(arg0 sim.Hook) {
+func (m *MockNamedHookable) AcceptHook(hook sim.Hook) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "AcceptHook", arg0)
+	m.ctrl.Call(m, "AcceptHook", hook)
 }
 
 // AcceptHook indicates an expected call of AcceptHook.
-func (mr *MockNamedHookableMockRecorder) AcceptHook(arg0 interface{}) *gomock.Call {
+func (mr *MockNamedHookableMockRecorder) AcceptHook(hook any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockNamedHookable)(nil).AcceptHook), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockNamedHookable)(nil).AcceptHook), hook)
 }
 
 // Hooks mocks base method.
@@ -66,7 +72,7 @@ func (m *MockNamedHookable) InvokeHook(arg0 sim.HookCtx) {
 }
 
 // InvokeHook indicates an expected call of InvokeHook.
-func (mr *MockNamedHookableMockRecorder) InvokeHook(arg0 interface{}) *gomock.Call {
+func (mr *MockNamedHookableMockRecorder) InvokeHook(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "InvokeHook", reflect.TypeOf((*MockNamedHookable)(nil).InvokeHook), arg0)
 }
