@@ -1,21 +1,27 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/sarchlab/mgpusim/v4/amd/timing/cp/internal/dispatching (interfaces: Dispatcher)
+//
+// Generated by this command:
+//
+//	mockgen -destination mock_dispatching_test.go -package cp -write_package_comment=false github.com/sarchlab/mgpusim/v4/amd/timing/cp/internal/dispatching Dispatcher
+//
 
 package cp
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	sim "github.com/sarchlab/akita/v4/sim"
 	protocol "github.com/sarchlab/mgpusim/v4/amd/protocol"
 	resource "github.com/sarchlab/mgpusim/v4/amd/timing/cp/internal/resource"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockDispatcher is a mock of Dispatcher interface.
 type MockDispatcher struct {
 	ctrl     *gomock.Controller
 	recorder *MockDispatcherMockRecorder
+	isgomock struct{}
 }
 
 // MockDispatcherMockRecorder is the mock recorder for MockDispatcher.
@@ -36,15 +42,15 @@ func (m *MockDispatcher) EXPECT() *MockDispatcherMockRecorder {
 }
 
 // AcceptHook mocks base method.
-func (m *MockDispatcher) AcceptHook(arg0 sim.Hook) {
+func (m *MockDispatcher) AcceptHook(hook sim.Hook) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "AcceptHook", arg0)
+	m.ctrl.Call(m, "AcceptHook", hook)
 }
 
 // AcceptHook indicates an expected call of AcceptHook.
-func (mr *MockDispatcherMockRecorder) AcceptHook(arg0 interface{}) *gomock.Call {
+func (mr *MockDispatcherMockRecorder) AcceptHook(hook any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockDispatcher)(nil).AcceptHook), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockDispatcher)(nil).AcceptHook), hook)
 }
 
 // Hooks mocks base method.
@@ -68,7 +74,7 @@ func (m *MockDispatcher) InvokeHook(arg0 sim.HookCtx) {
 }
 
 // InvokeHook indicates an expected call of InvokeHook.
-func (mr *MockDispatcherMockRecorder) InvokeHook(arg0 interface{}) *gomock.Call {
+func (mr *MockDispatcherMockRecorder) InvokeHook(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "InvokeHook", reflect.TypeOf((*MockDispatcher)(nil).InvokeHook), arg0)
 }
@@ -116,27 +122,27 @@ func (mr *MockDispatcherMockRecorder) NumHooks() *gomock.Call {
 }
 
 // RegisterCU mocks base method.
-func (m *MockDispatcher) RegisterCU(arg0 resource.DispatchableCU) {
+func (m *MockDispatcher) RegisterCU(cu resource.DispatchableCU) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "RegisterCU", arg0)
+	m.ctrl.Call(m, "RegisterCU", cu)
 }
 
 // RegisterCU indicates an expected call of RegisterCU.
-func (mr *MockDispatcherMockRecorder) RegisterCU(arg0 interface{}) *gomock.Call {
+func (mr *MockDispatcherMockRecorder) RegisterCU(cu any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RegisterCU", reflect.TypeOf((*MockDispatcher)(nil).RegisterCU), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RegisterCU", reflect.TypeOf((*MockDispatcher)(nil).RegisterCU), cu)
 }
 
 // StartDispatching mocks base method.
-func (m *MockDispatcher) StartDispatching(arg0 *protocol.LaunchKernelReq) {
+func (m *MockDispatcher) StartDispatching(req *protocol.LaunchKernelReq) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "StartDispatching", arg0)
+	m.ctrl.Call(m, "StartDispatching", req)
 }
 
 // StartDispatching indicates an expected call of StartDispatching.
-func (mr *MockDispatcherMockRecorder) StartDispatching(arg0 interface{}) *gomock.Call {
+func (mr *MockDispatcherMockRecorder) StartDispatching(req any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "StartDispatching", reflect.TypeOf((*MockDispatcher)(nil).StartDispatching), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "StartDispatching", reflect.TypeOf((*MockDispatcher)(nil).StartDispatching), req)
 }
 
 // Tick mocks base method.
