@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: coalescer.go
+//
+// Generated by this command:
+//
+//	mockgen -source coalescer.go -destination mock_coalsecer_test.go -package cu
+//
 
 // Package cu is a generated GoMock package.
 package cu
@@ -7,14 +12,15 @@ package cu
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	wavefront "github.com/sarchlab/mgpusim/v4/amd/timing/wavefront"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // Mockcoalescer is a mock of coalescer interface.
 type Mockcoalescer struct {
 	ctrl     *gomock.Controller
 	recorder *MockcoalescerMockRecorder
+	isgomock struct{}
 }
 
 // MockcoalescerMockRecorder is the mock recorder for Mockcoalescer.
@@ -43,7 +49,7 @@ func (m *Mockcoalescer) generateMemTransactions(wf *wavefront.Wavefront) []Vecto
 }
 
 // generateMemTransactions indicates an expected call of generateMemTransactions.
-func (mr *MockcoalescerMockRecorder) generateMemTransactions(wf interface{}) *gomock.Call {
+func (mr *MockcoalescerMockRecorder) generateMemTransactions(wf any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "generateMemTransactions", reflect.TypeOf((*Mockcoalescer)(nil).generateMemTransactions), wf)
 }
