@@ -1,20 +1,26 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/sarchlab/akita/v4/pipelining (interfaces: Pipeline)
+//
+// Generated by this command:
+//
+//	mockgen -write_package_comment=false -package=cu -destination=mock_pipelining_test.go github.com/sarchlab/akita/v4/pipelining Pipeline
+//
 
 package cu
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	pipelining "github.com/sarchlab/akita/v4/pipelining"
 	sim "github.com/sarchlab/akita/v4/sim"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockPipeline is a mock of Pipeline interface.
 type MockPipeline struct {
 	ctrl     *gomock.Controller
 	recorder *MockPipelineMockRecorder
+	isgomock struct{}
 }
 
 // MockPipelineMockRecorder is the mock recorder for MockPipeline.
@@ -35,27 +41,27 @@ func (m *MockPipeline) EXPECT() *MockPipelineMockRecorder {
 }
 
 // Accept mocks base method.
-func (m *MockPipeline) Accept(arg0 pipelining.PipelineItem) {
+func (m *MockPipeline) Accept(elem pipelining.PipelineItem) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "Accept", arg0)
+	m.ctrl.Call(m, "Accept", elem)
 }
 
 // Accept indicates an expected call of Accept.
-func (mr *MockPipelineMockRecorder) Accept(arg0 interface{}) *gomock.Call {
+func (mr *MockPipelineMockRecorder) Accept(elem any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Accept", reflect.TypeOf((*MockPipeline)(nil).Accept), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Accept", reflect.TypeOf((*MockPipeline)(nil).Accept), elem)
 }
 
 // AcceptHook mocks base method.
-func (m *MockPipeline) AcceptHook(arg0 sim.Hook) {
+func (m *MockPipeline) AcceptHook(hook sim.Hook) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "AcceptHook", arg0)
+	m.ctrl.Call(m, "AcceptHook", hook)
 }
 
 // AcceptHook indicates an expected call of AcceptHook.
-func (mr *MockPipelineMockRecorder) AcceptHook(arg0 interface{}) *gomock.Call {
+func (mr *MockPipelineMockRecorder) AcceptHook(hook any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockPipeline)(nil).AcceptHook), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockPipeline)(nil).AcceptHook), hook)
 }
 
 // CanAccept mocks base method.
@@ -105,7 +111,7 @@ func (m *MockPipeline) InvokeHook(arg0 sim.HookCtx) {
 }
 
 // InvokeHook indicates an expected call of InvokeHook.
-func (mr *MockPipelineMockRecorder) InvokeHook(arg0 interface{}) *gomock.Call {
+func (mr *MockPipelineMockRecorder) InvokeHook(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "InvokeHook", reflect.TypeOf((*MockPipeline)(nil).InvokeHook), arg0)
 }
