@@ -1,19 +1,25 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/sarchlab/akita/v4/sim (interfaces: Port,Engine,Buffer)
+//
+// Generated by this command:
+//
+//	mockgen -write_package_comment=false -package=cu -destination=mock_sim_test.go github.com/sarchlab/akita/v4/sim Port,Engine,Buffer
+//
 
 package cu
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	sim "github.com/sarchlab/akita/v4/sim"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockPort is a mock of Port interface.
 type MockPort struct {
 	ctrl     *gomock.Controller
 	recorder *MockPortMockRecorder
+	isgomock struct{}
 }
 
 // MockPortMockRecorder is the mock recorder for MockPort.
@@ -34,15 +40,15 @@ func (m *MockPort) EXPECT() *MockPortMockRecorder {
 }
 
 // AcceptHook mocks base method.
-func (m *MockPort) AcceptHook(arg0 sim.Hook) {
+func (m *MockPort) AcceptHook(hook sim.Hook) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "AcceptHook", arg0)
+	m.ctrl.Call(m, "AcceptHook", hook)
 }
 
 // AcceptHook indicates an expected call of AcceptHook.
-func (mr *MockPortMockRecorder) AcceptHook(arg0 interface{}) *gomock.Call {
+func (mr *MockPortMockRecorder) AcceptHook(hook any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockPort)(nil).AcceptHook), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockPort)(nil).AcceptHook), hook)
 }
 
 // AsRemote mocks base method.
@@ -88,17 +94,17 @@ func (mr *MockPortMockRecorder) Component() *gomock.Call {
 }
 
 // Deliver mocks base method.
-func (m *MockPort) Deliver(arg0 sim.Msg) *sim.SendError {
+func (m *MockPort) Deliver(msg sim.Msg) *sim.SendError {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Deliver", arg0)
+	ret := m.ctrl.Call(m, "Deliver", msg)
 	ret0, _ := ret[0].(*sim.SendError)
 	return ret0
 }
 
 // Deliver indicates an expected call of Deliver.
-func (mr *MockPortMockRecorder) Deliver(arg0 interface{}) *gomock.Call {
+func (mr *MockPortMockRecorder) Deliver(msg any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Deliver", reflect.TypeOf((*MockPort)(nil).Deliver), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Deliver", reflect.TypeOf((*MockPort)(nil).Deliver), msg)
 }
 
 // Hooks mocks base method.
@@ -212,35 +218,36 @@ func (mr *MockPortMockRecorder) RetrieveOutgoing() *gomock.Call {
 }
 
 // Send mocks base method.
-func (m *MockPort) Send(arg0 sim.Msg) *sim.SendError {
+func (m *MockPort) Send(msg sim.Msg) *sim.SendError {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Send", arg0)
+	ret := m.ctrl.Call(m, "Send", msg)
 	ret0, _ := ret[0].(*sim.SendError)
 	return ret0
 }
 
 // Send indicates an expected call of Send.
-func (mr *MockPortMockRecorder) Send(arg0 interface{}) *gomock.Call {
+func (mr *MockPortMockRecorder) Send(msg any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Send", reflect.TypeOf((*MockPort)(nil).Send), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Send", reflect.TypeOf((*MockPort)(nil).Send), msg)
 }
 
 // SetConnection mocks base method.
-func (m *MockPort) SetConnection(arg0 sim.Connection) {
+func (m *MockPort) SetConnection(conn sim.Connection) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "SetConnection", arg0)
+	m.ctrl.Call(m, "SetConnection", conn)
 }
 
 // SetConnection indicates an expected call of SetConnection.
-func (mr *MockPortMockRecorder) SetConnection(arg0 interface{}) *gomock.Call {
+func (mr *MockPortMockRecorder) SetConnection(conn any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetConnection", reflect.TypeOf((*MockPort)(nil).SetConnection), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetConnection", reflect.TypeOf((*MockPort)(nil).SetConnection), conn)
 }
 
 // MockEngine is a mock of Engine interface.
 type MockEngine struct {
 	ctrl     *gomock.Controller
 	recorder *MockEngineMockRecorder
+	isgomock struct{}
 }
 
 // MockEngineMockRecorder is the mock recorder for MockEngine.
@@ -261,15 +268,15 @@ func (m *MockEngine) EXPECT() *MockEngineMockRecorder {
 }
 
 // AcceptHook mocks base method.
-func (m *MockEngine) AcceptHook(arg0 sim.Hook) {
+func (m *MockEngine) AcceptHook(hook sim.Hook) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "AcceptHook", arg0)
+	m.ctrl.Call(m, "AcceptHook", hook)
 }
 
 // AcceptHook indicates an expected call of AcceptHook.
-func (mr *MockEngineMockRecorder) AcceptHook(arg0 interface{}) *gomock.Call {
+func (mr *MockEngineMockRecorder) AcceptHook(hook any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockEngine)(nil).AcceptHook), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockEngine)(nil).AcceptHook), hook)
 }
 
 // Continue mocks base method.
@@ -353,21 +360,22 @@ func (mr *MockEngineMockRecorder) Run() *gomock.Call {
 }
 
 // Schedule mocks base method.
-func (m *MockEngine) Schedule(arg0 sim.Event) {
+func (m *MockEngine) Schedule(e sim.Event) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "Schedule", arg0)
+	m.ctrl.Call(m, "Schedule", e)
 }
 
 // Schedule indicates an expected call of Schedule.
-func (mr *MockEngineMockRecorder) Schedule(arg0 interface{}) *gomock.Call {
+func (mr *MockEngineMockRecorder) Schedule(e any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Schedule", reflect.TypeOf((*MockEngine)(nil).Schedule), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Schedule", reflect.TypeOf((*MockEngine)(nil).Schedule), e)
 }
 
 // MockBuffer is a mock of Buffer interface.
 type MockBuffer struct {
 	ctrl     *gomock.Controller
 	recorder *MockBufferMockRecorder
+	isgomock struct{}
 }
 
 // MockBufferMockRecorder is the mock recorder for MockBuffer.
@@ -388,15 +396,15 @@ func (m *MockBuffer) EXPECT() *MockBufferMockRecorder {
 }
 
 // AcceptHook mocks base method.
-func (m *MockBuffer) AcceptHook(arg0 sim.Hook) {
+func (m *MockBuffer) AcceptHook(hook sim.Hook) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "AcceptHook", arg0)
+	m.ctrl.Call(m, "AcceptHook", hook)
 }
 
 // AcceptHook indicates an expected call of AcceptHook.
-func (mr *MockBufferMockRecorder) AcceptHook(arg0 interface{}) *gomock.Call {
+func (mr *MockBufferMockRecorder) AcceptHook(hook any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockBuffer)(nil).AcceptHook), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptHook", reflect.TypeOf((*MockBuffer)(nil).AcceptHook), hook)
 }
 
 // CanPush mocks base method.
@@ -482,10 +490,10 @@ func (mr *MockBufferMockRecorder) NumHooks() *gomock.Call {
 }
 
 // Peek mocks base method.
-func (m *MockBuffer) Peek() interface{} {
+func (m *MockBuffer) Peek() any {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Peek")
-	ret0, _ := ret[0].(interface{})
+	ret0, _ := ret[0].(any)
 	return ret0
 }
 
@@ -496,10 +504,10 @@ func (mr *MockBufferMockRecorder) Peek() *gomock.Call {
 }
 
 // Pop mocks base method.
-func (m *MockBuffer) Pop() interface{} {
+func (m *MockBuffer) Pop() any {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Pop")
-	ret0, _ := ret[0].(interface{})
+	ret0, _ := ret[0].(any)
 	return ret0
 }
 
@@ -510,15 +518,15 @@ func (mr *MockBufferMockRecorder) Pop() *gomock.Call {
 }
 
 // Push mocks base method.
-func (m *MockBuffer) Push(arg0 interface{}) {
+func (m *MockBuffer) Push(e any) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "Push", arg0)
+	m.ctrl.Call(m, "Push", e)
 }
 
 // Push indicates an expected call of Push.
-func (mr *MockBufferMockRecorder) Push(arg0 interface{}) *gomock.Call {
+func (mr *MockBufferMockRecorder) Push(e any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Push", reflect.TypeOf((*MockBuffer)(nil).Push), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Push", reflect.TypeOf((*MockBuffer)(nil).Push), e)
 }
 
 // Size mocks base method.
