@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: subcomponent.go
+//
+// Generated by this command:
+//
+//	mockgen -source subcomponent.go -destination mock_subcomponent_test.go -package cu
+//
 
 // Package cu is a generated GoMock package.
 package cu
@@ -7,14 +12,15 @@ package cu
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	wavefront "github.com/sarchlab/mgpusim/v4/amd/timing/wavefront"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockSubComponent is a mock of SubComponent interface.
 type MockSubComponent struct {
 	ctrl     *gomock.Controller
 	recorder *MockSubComponentMockRecorder
+	isgomock struct{}
 }
 
 // MockSubComponentMockRecorder is the mock recorder for MockSubComponent.
@@ -41,7 +47,7 @@ func (m *MockSubComponent) AcceptWave(wave *wavefront.Wavefront) {
 }
 
 // AcceptWave indicates an expected call of AcceptWave.
-func (mr *MockSubComponentMockRecorder) AcceptWave(wave interface{}) *gomock.Call {
+func (mr *MockSubComponentMockRecorder) AcceptWave(wave any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AcceptWave", reflect.TypeOf((*MockSubComponent)(nil).AcceptWave), wave)
 }
