@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: wfdispatcher.go
+//
+// Generated by this command:
+//
+//	mockgen -source wfdispatcher.go -destination mock_wfdispatcher_test.go -package cu
+//
 
 // Package cu is a generated GoMock package.
 package cu
@@ -7,15 +12,16 @@ package cu
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	protocol "github.com/sarchlab/mgpusim/v4/amd/protocol"
 	wavefront "github.com/sarchlab/mgpusim/v4/amd/timing/wavefront"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockWfDispatcher is a mock of WfDispatcher interface.
 type MockWfDispatcher struct {
 	ctrl     *gomock.Controller
 	recorder *MockWfDispatcherMockRecorder
+	isgomock struct{}
 }
 
 // MockWfDispatcherMockRecorder is the mock recorder for MockWfDispatcher.
@@ -42,7 +48,7 @@ func (m *MockWfDispatcher) DispatchWf(wf *wavefront.Wavefront, location protocol
 }
 
 // DispatchWf indicates an expected call of DispatchWf.
-func (mr *MockWfDispatcherMockRecorder) DispatchWf(wf, location interface{}) *gomock.Call {
+func (mr *MockWfDispatcherMockRecorder) DispatchWf(wf, location any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DispatchWf", reflect.TypeOf((*MockWfDispatcher)(nil).DispatchWf), wf, location)
 }
