@@ -5,13 +5,15 @@ import (
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/sarchlab/akita/v4/simulation"
 	"github.com/sarchlab/mgpusim/v4/amd/benchmarks/mccl"
 	"github.com/sarchlab/mgpusim/v4/amd/driver"
-	"github.com/sarchlab/mgpusim/v4/amd/samples/runner"
+	"github.com/sarchlab/mgpusim/v4/amd/samples/runner/timingconfig"
 )
 
 var _ = Describe("MCCL", func() {
 	var (
+		s         *simulation.Simulation
 		gpuDriver *driver.Driver
 		context   *driver.Context
 		gpuIDs    []int
@@ -19,16 +21,18 @@ var _ = Describe("MCCL", func() {
 	)
 
 	BeforeEach(func() {
-		platform := runner.MakeR9NanoBuilder().
-			WithNumGPU(4).
+		s = simulation.MakeBuilder().WithoutMonitoring().Build()
+		timingconfig.MakeBuilder().
+			WithSimulation(s).
+			WithNumGPUs(4).
 			Build()
-		gpuDriver = platform.Driver
+		gpuDriver = s.GetComponentByName("Driver").(*driver.Driver)
 		gpuDriver.Run()
 		context = gpuDriver.Init()
 	})
 
 	AfterEach(func() {
-		gpuDriver.Terminate()
+		s.Terminate()
 	})
 
 	It("Broadcast Test", func() {
