@@ -1,10 +1,10 @@
 package driver
 
 import (
-	"github.com/golang/mock/gomock"
 	"github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/sarchlab/akita/v4/mem/vm"
+	"go.uber.org/mock/gomock"
 )
 
 var _ = ginkgo.Describe("Distributor", func() {
