@@ -1,12 +1,12 @@
 package emu
 
 import (
-	"github.com/golang/mock/gomock"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/sarchlab/akita/v4/mem/mem"
 	"github.com/sarchlab/akita/v4/mem/vm"
 	"github.com/sarchlab/mgpusim/v4/amd/insts"
+	"go.uber.org/mock/gomock"
 )
 
 type mockInstState struct {
