@@ -6,18 +6,16 @@ import (
 
 	// Enable profiling
 	_ "net/http/pprof"
-	"strconv"
-	"strings"
 	"sync"
 
-	"github.com/sarchlab/akita/v4/monitoring"
 	"github.com/sarchlab/akita/v4/sim"
+	"github.com/sarchlab/akita/v4/simulation"
 	"github.com/sarchlab/akita/v4/tracing"
 	"github.com/sarchlab/mgpusim/v4/amd/benchmarks"
 	"github.com/sarchlab/mgpusim/v4/amd/driver"
+	"github.com/sarchlab/mgpusim/v4/amd/samples/runner/emusystem"
+	"github.com/sarchlab/mgpusim/v4/amd/samples/runner/timingconfig"
 	"github.com/sarchlab/mgpusim/v4/amd/sampling"
-
-	"github.com/tebeka/atexit"
 )
 
 type verificationPreEnablingBenchmark interface {
@@ -28,45 +26,26 @@ type verificationPreEnablingBenchmark interface {
 
 // Runner is a class that helps running the benchmarks in the official samples.
 type Runner struct {
-	platform                *Platform
-	maxInstStopper          *instTracer
-	kernelTimeCounter       *tracing.BusyTimeTracer
-	perGPUKernelTimeCounter []*tracing.BusyTimeTracer
-	instCountTracers        []instCountTracer
-	cacheLatencyTracers     []cacheLatencyTracer
-	cacheHitRateTracers     []cacheHitRateTracer
-	tlbHitRateTracers       []tlbHitRateTracer
-	rdmaTransactionCounters []rdmaTransactionCountTracer
-	dramTracers             []dramTransactionCountTracer
-	benchmarks              []benchmarks.Benchmark
-	monitor                 *monitoring.Monitor
-	metricsCollector        *collector
-	simdBusyTimeTracers     []simdBusyTimeTracer
-	cuCPITraces             []cuCPIStackTracer
+	simulation *simulation.Simulation
+	platform   *sim.Domain
+	reporter   *reporter
 
-	Timing                     bool
-	Verify                     bool
-	Parallel                   bool
-	ReportInstCount            bool
-	ReportCacheLatency         bool
-	ReportCacheHitRate         bool
-	ReportTLBHitRate           bool
-	ReportRDMATransactionCount bool
-	ReportDRAMTransactionCount bool
-	UseUnifiedMemory           bool
-	ReportSIMDBusyTime         bool
-	ReportCPIStack             bool
+	Timing           bool
+	Verify           bool
+	Parallel         bool
+	UseUnifiedMemory bool
 
-	GPUIDs []int
+	GPUIDs     []int
+	benchmarks []benchmarks.Benchmark
 }
 
 // Init initializes the platform simulate
 func (r *Runner) Init() *Runner {
-	r.ParseFlag()
-	r.parseGPUFlag()
+	r.parseFlag()
 
 	log.SetFlags(log.Llongfile | log.Ldate | log.Ltime)
-	sampling.InitSampledEngine()
+
+	r.initSimulation()
 
 	if r.Timing {
 		r.buildTimingPlatform()
@@ -76,128 +55,56 @@ func (r *Runner) Init() *Runner {
 
 	r.createUnifiedGPUs()
 
-	r.defineMetrics()
-
 	return r
 }
 
-func (r *Runner) buildEmuPlatform() {
-	b := MakeEmuBuilder().
-		WithNumGPU(r.GPUIDs[len(r.GPUIDs)-1])
+func (r *Runner) initSimulation() {
+	builder := simulation.MakeBuilder()
 
-	if r.Parallel {
-		b = b.WithParallelEngine()
+	if *parallelFlag {
+		builder = builder.WithParallelEngine()
 	}
+
+	r.simulation = builder.Build()
+}
+
+func (r *Runner) buildEmuPlatform() {
+	b := emusystem.MakeBuilder().
+		WithSimulation(r.simulation).
+		WithNumGPUs(r.GPUIDs[len(r.GPUIDs)-1])
 
 	if *isaDebug {
-		b = b.WithISADebugging()
-	}
-
-	if *visTracing {
-		b = b.WithVisTracing()
-	}
-
-	if *memTracing {
-		b = b.WithMemTracing()
-	}
-
-	if *magicMemoryCopy {
-		b = b.WithMagicMemoryCopy()
+		b = b.WithDebugISA()
 	}
 
 	r.platform = b.Build()
 }
 
 func (r *Runner) buildTimingPlatform() {
-	b := MakeR9NanoBuilder().
-		WithNumGPU(r.GPUIDs[len(r.GPUIDs)-1])
+	sampling.InitSampledEngine()
 
-	if r.Parallel {
-		b = b.WithParallelEngine()
-	}
-
-	if *isaDebug {
-		b = b.WithISADebugging()
-	}
-
-	if *visTracing {
-		b = b.WithPartialVisTracing(
-			sim.VTimeInSec(*visTraceStartTime),
-			sim.VTimeInSec(*visTraceEndTime),
-		)
-	}
-
-	if *memTracing {
-		b = b.WithMemTracing()
-	}
-
-	r.monitor = monitoring.NewMonitor()
-	if *customPortForAkitaRTM != 0 {
-		r.monitor = r.monitor.WithPortNumber(*customPortForAkitaRTM)
-	}
-
-	b = b.WithMonitor(r.monitor)
-
-	b = r.setAnalyzer(b)
+	b := timingconfig.MakeBuilder().
+		WithSimulation(r.simulation).
+		WithNumGPUs(r.GPUIDs[len(r.GPUIDs)-1])
 
 	if *magicMemoryCopy {
 		b = b.WithMagicMemoryCopy()
 	}
 
 	r.platform = b.Build()
-
-	if !*disableAkitaRTM {
-		r.monitor.StartServer()
-	}
+	r.reporter = newReporter(r.simulation)
+	r.configureVisTracing()
 }
 
-func (*Runner) setAnalyzer(
-	b R9NanoPlatformBuilder,
-) R9NanoPlatformBuilder {
-	if *analyzerPeriodFlag != 0 && *analyzerNameFlag == "" {
-		panic("must specify -analyzer-name when using -analyzer-period")
-	}
-
-	if *analyzerNameFlag != "" {
-		b = b.WithPerfAnalyzer(
-			*analyzerNameFlag,
-			*analyzerPeriodFlag,
-		)
-	}
-	return b
-}
-
-func (r *Runner) addMaxInstStopper() {
-	if *maxInstCount == 0 {
+func (r *Runner) configureVisTracing() {
+	if !*visTracing {
 		return
 	}
 
-	r.maxInstStopper = newInstStopper(*maxInstCount)
-	for _, gpu := range r.platform.GPUs {
-		for _, cu := range gpu.CUs {
-			tracing.CollectTrace(cu.(tracing.NamedHookable), r.maxInstStopper)
-		}
+	visTracer := r.simulation.GetVisTracer()
+	for _, comp := range r.simulation.Components() {
+		tracing.CollectTrace(comp.(tracing.NamedHookable), visTracer)
 	}
-}
-
-func (r *Runner) parseGPUFlag() {
-	if *gpuFlag == "" && *unifiedGPUFlag == "" {
-		r.GPUIDs = []int{1}
-		return
-	}
-
-	if *gpuFlag != "" && *unifiedGPUFlag != "" {
-		panic("cannot use -gpus and -unified-gpus together")
-	}
-
-	var gpuIDs []int
-	if *gpuFlag != "" {
-		gpuIDs = r.gpuIDStringToList(*gpuFlag)
-	} else if *unifiedGPUFlag != "" {
-		gpuIDs = r.gpuIDStringToList(*unifiedGPUFlag)
-	}
-
-	r.GPUIDs = gpuIDs
 }
 
 func (r *Runner) createUnifiedGPUs() {
@@ -205,23 +112,9 @@ func (r *Runner) createUnifiedGPUs() {
 		return
 	}
 
-	unifiedGPUID := r.platform.Driver.CreateUnifiedGPU(nil, r.GPUIDs)
+	driver := r.simulation.GetComponentByName("Driver").(*driver.Driver)
+	unifiedGPUID := driver.CreateUnifiedGPU(nil, r.GPUIDs)
 	r.GPUIDs = []int{unifiedGPUID}
-}
-
-func (r *Runner) gpuIDStringToList(gpuIDsString string) []int {
-	gpuIDs := make([]int, 0)
-	gpuIDTokens := strings.Split(gpuIDsString, ",")
-
-	for _, t := range gpuIDTokens {
-		gpuID, err := strconv.Atoi(t)
-		if err != nil {
-			panic(err)
-		}
-		gpuIDs = append(gpuIDs, gpuID)
-	}
-
-	return gpuIDs
 }
 
 // AddBenchmark adds an benchmark that the driver runs
@@ -230,6 +123,7 @@ func (r *Runner) AddBenchmark(b benchmarks.Benchmark) {
 	if r.UseUnifiedMemory {
 		b.SetUnifiedMemory()
 	}
+
 	r.benchmarks = append(r.benchmarks, b)
 }
 
@@ -239,12 +133,13 @@ func (r *Runner) AddBenchmarkWithoutSettingGPUsToUse(b benchmarks.Benchmark) {
 	if r.UseUnifiedMemory {
 		b.SetUnifiedMemory()
 	}
+
 	r.benchmarks = append(r.benchmarks, b)
 }
 
-// Run runs the benchmark on the simulator
+// Run runs the benchmark
 func (r *Runner) Run() {
-	r.platform.Driver.Run()
+	r.Driver().Run()
 
 	var wg sync.WaitGroup
 	for _, b := range r.benchmarks {
@@ -266,17 +161,20 @@ func (r *Runner) Run() {
 	}
 	wg.Wait()
 
-	r.platform.Driver.Terminate()
+	if r.reporter != nil {
+		r.reporter.report()
+	}
 
-	atexit.Exit(0)
+	r.Driver().Terminate()
+	r.simulation.Terminate()
 }
 
 // Driver returns the GPU driver used by the current runner.
 func (r *Runner) Driver() *driver.Driver {
-	return r.platform.Driver
+	return r.simulation.GetComponentByName("Driver").(*driver.Driver)
 }
 
 // Engine returns the event-driven simulation engine used by the current runner.
 func (r *Runner) Engine() sim.Engine {
-	return r.platform.Engine
+	return r.simulation.GetEngine()
 }
