@@ -1,12 +1,12 @@
 package cp
 
 import (
-	"github.com/golang/mock/gomock"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/sarchlab/akita/v4/mem/mem"
 	"github.com/sarchlab/akita/v4/sim"
 	"github.com/sarchlab/mgpusim/v4/amd/protocol"
+	"go.uber.org/mock/gomock"
 )
 
 var _ = Describe("DMAEngine", func() {
