@@ -1,7 +1,6 @@
 package cu
 
 import (
-	"github.com/golang/mock/gomock"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/sarchlab/akita/v4/mem/mem"
@@ -9,6 +8,7 @@ import (
 	"github.com/sarchlab/mgpusim/v4/amd/emu"
 	"github.com/sarchlab/mgpusim/v4/amd/insts"
 	"github.com/sarchlab/mgpusim/v4/amd/timing/wavefront"
+	"go.uber.org/mock/gomock"
 )
 
 type mockScratchpadPreparer struct {
