@@ -1,13 +1,13 @@
 package cu
 
 import (
-	"github.com/golang/mock/gomock"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/sarchlab/akita/v4/mem/mem"
 	"github.com/sarchlab/mgpusim/v4/amd/insts"
 	"github.com/sarchlab/mgpusim/v4/amd/kernels"
 	"github.com/sarchlab/mgpusim/v4/amd/timing/wavefront"
+	"go.uber.org/mock/gomock"
 )
 
 var _ = Describe("Vector Memory Unit", func() {
